perf-test/requirements.txt
locust-plugins
-e ../metadata-ingestion[datahub-rest,datahub-kafka]
kubernetes